refactor trees and folders
